internal/randvar: take uint64 in IncMax
Take a uint64, not an int, in (randvar.Dynamic).IncMax to match the interface's
other methods.

+++ cmd/pebble/ycsb.go
@@ -482,7 +482,7 @@ func (y *ycsb) insert(db DB, buf *ycsbBuf) {
 			log.Fatal(err)
 		}
 		if delta > 0 {
-			y.keyDist.IncMax(delta)
+			y.keyDist.IncMax(uint64(delta))
 		}
 	}
 }
